+++ NEWS.md
@@ -1,3 +1,9 @@
+# DataFrames.jl v1.5 Release Notes
+
+* add `allunique` and allow transformations in `cols` argument of `describe`
+  and `nonunique` when working with `SubDataFrame`
+  ([3232](https://github.com/JuliaData/DataFrames.jl/pull/3232))
+
 # DataFrames.jl v1.4.3 Patch Release Notes
 
 ## Bug fixes